move CloudProvider to service.cloud package
