+++ server/odc-service/src/main/java/com/oceanbase/odc/service/objectstorage/cloud/model/CloudObjectStorageProperties.java
@@ -19,7 +19,7 @@ import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.cloud.context.config.annotation.RefreshScope;
 import org.springframework.context.annotation.Configuration;
 
-import com.oceanbase.odc.service.objectstorage.cloud.model.ObjectStorageConfiguration.CloudProvider;
+import com.oceanbase.odc.service.cloud.model.CloudProvider;
 
 import lombok.Data;
